[#250] refactor: search, mainpage 패키지 구조 이동
