+++ linkmind/src/main/java/com/app/toaster/mainpage/controller/MainController.java
@@ -1,11 +1,11 @@
-package com.app.toaster.controller;
+package com.app.toaster.mainpage.controller;
 
 import com.app.toaster.config.UserId;
 import org.springframework.web.bind.annotation.*;
 
 import com.app.toaster.common.dto.ApiResponse;
 import com.app.toaster.controller.valid.TitleValid;
-import com.app.toaster.service.search.SearchService;
+import com.app.toaster.search.service.SearchService;
 
 import lombok.RequiredArgsConstructor;
