+++ linkmind/src/main/java/com/app/toaster/search/controller/response/CategoryResult.java
@@ -1,4 +1,4 @@
-package com.app.toaster.controller.response.search;
+package com.app.toaster.search.controller.response;
 
 public record CategoryResult(Long categoryId, String title, Long toastNum) {
 	public static CategoryResult of(Long categoryId, String title, Long toastNum){
